Decoupled lib to better maintainability
