+++ src/index.js
@@ -1,7 +1,9 @@
 "use strict";
 
 const services = require('./utils/services');
-const { loadDomain, processOperation, StrategiesType } = require('./utils/index');
+const { loadDomain, StrategiesType } = require('./utils/index');
+const { Key, searchBypassed } = require('./lib/bypasser');
+const checkCriteriaOffline = require('./lib/resolver');
 const fs = require('fs');
 
 const DEFAULT_SNAPSHOT_LOCATION = './snapshot/';
@@ -163,93 +165,6 @@ class Switcher {
     return StrategiesType;
   }
   
-}
-
-class Key {
-  constructor(key) {
-    this.key = key;
-    this.value = undefined;
-  }
-
-  true() {
-    this.value = true;
-  }
-
-  false() {
-    this.value = false;
-  }
-
-  getKey() {
-    return this.key;
-  }
-
-  getValue() {
-    return this.value;
-  }
-}
-
-function searchBypassed(key, bypassedKeys) {
-  let existentKey;
-  bypassedKeys.forEach(async bk => {
-    if (bk.getKey() === key) {
-      return existentKey = bk;
-    }
-  })
-  return existentKey;
-}
-
-async function checkCriteriaOffline(key, input, snapshot) {
-  const { data } = snapshot;
-  return await resolveCriteria(key, input, data);
-}
-
-async function resolveCriteria(key, input, { domain }) {
-  if (!domain.activated) {
-    return false;
-  }
-
-  let result = true;
-  let notFoundKey = true;
-  const { group } = domain;
-  if (group) {
-    group.forEach(function (g) {
-      
-      const { config } = g;
-      const configFound = config.filter(c => c.key === key)
-
-      if (configFound[0]) {
-        notFoundKey = false;
-        if (!g.activated) {
-          return result = false;
-        }
-
-        if (!configFound[0].activated) {
-          return result = false;
-        }
-        
-        const { strategies } = configFound[0]
-        strategies.forEach(function(strategy) {
-          if (strategy.activated) {
-            if (!input) {
-              return result = false;
-            }
-
-            const entry = services.getEntry(input);
-            const entryInput = entry.filter(e => e.strategy === strategy.strategy)
-            if (!processOperation(strategy.strategy, strategy.operation, entryInput[0].input, strategy.values)) {
-              return result = false;
-            }
-          }
-        })
-      }
-    })
-  }
-
-  if (notFoundKey) {
-    throw new Error(`Something went wrong: {"error":"Unable to load a key ${key}"}`)
-  }
-
-  return result;
 }
 
 module.exports = Switcher;
